+++ src/unit_tests/wazuh_modules/CMakeLists.txt
@@ -18,6 +18,7 @@ if(${TARGET} STREQUAL "server")
   add_subdirectory(sca)
   add_subdirectory(scheduling)
   add_subdirectory(vulnerability_detector)
+  add_subdirectory(task_manager)
 elseif(${TARGET} STREQUAL "agent")
   add_subdirectory(syscollector)
 endif()
