+++ src/wazuh_modules/task_manager/wm_task_manager_db.c
@@ -28,32 +28,15 @@ pthread_mutex_t db_mutex = PTHREAD_MUTEX_INITIALIZER;
 
 STATIC int wm_task_manager_sql_error(sqlite3 *db, sqlite3_stmt *stmt);
 
-/**
- * Update old tasks with status in progress to status timeout
- * @param now Actual time
- * @param timeout Task timeout
- * @param next_timeout Next task in progress timeout
- * @return OS_SUCCESS on success, OS_INVALID on errors
- * */
-STATIC int wm_task_manager_set_timeout_status(time_t now, int timeout, time_t *next_timeout) __attribute__((nonnull));
-
-/**
- * Delete old tasks from the tasks DB
- * @param timestamp Deletion limit time
- * @return OS_SUCCESS on success, OS_INVALID on errors
- * */
-STATIC int wm_task_manager_delete_old_entries(int timestamp);
-
 static const char *task_queries[] = {
     [WM_TASK_INSERT_TASK] = "INSERT INTO " TASKS_TABLE " VALUES(NULL,?,?,?,?,?,?,?,?);",
     [WM_TASK_GET_LAST_AGENT_TASK] = "SELECT *, MAX(CREATE_TIME) FROM " TASKS_TABLE " WHERE AGENT_ID = ?;",
     [WM_TASK_GET_LAST_AGENT_UPGRADE_TASK] = "SELECT *, MAX(CREATE_TIME) FROM " TASKS_TABLE " WHERE AGENT_ID = ? AND (COMMAND = 'upgrade' OR COMMAND = 'upgrade_custom');",
     [WM_TASK_UPDATE_TASK_STATUS] = "UPDATE " TASKS_TABLE " SET STATUS = ?, LAST_UPDATE_TIME = ?, ERROR_MESSAGE = ? WHERE TASK_ID = ?;",
-    [WM_TASK_GET_TASK_BY_TASK_ID] = "SELECT * FROM " TASKS_TABLE " WHERE TASK_ID = ?;",
     [WM_TASK_GET_TASK_BY_STATUS] = "SELECT * FROM " TASKS_TABLE " WHERE STATUS = ?;",
     [WM_TASK_DELETE_OLD_TASKS] = "DELETE FROM " TASKS_TABLE " WHERE CREATE_TIME <= ?;",
     [WM_TASK_DELETE_TASK] = "DELETE FROM " TASKS_TABLE " WHERE TASK_ID = ?;",
-    [WM_TASK_CANCEL_PENDING_UPGRADE_TASKS] = "UPDATE " TASKS_TABLE " SET STATUS = '" WM_TASK_STATUS_CANCELLED "', LAST_UPDATE_TIME = ? WHERE NODE = ? AND STATUS = '" WM_TASK_STATUS_PENDING "';"
+    [WM_TASK_CANCEL_PENDING_UPGRADE_TASKS] = "UPDATE " TASKS_TABLE " SET STATUS = '" WM_TASK_STATUS_CANCELLED "', LAST_UPDATE_TIME = ? WHERE NODE = ? AND STATUS = '" WM_TASK_STATUS_PENDING "' AND (COMMAND = 'upgrade' OR COMMAND = 'upgrade_custom');"
 };
 
 STATIC int wm_task_manager_sql_error(sqlite3 *db, sqlite3_stmt *stmt) {
@@ -63,109 +46,7 @@ STATIC int wm_task_manager_sql_error(sqlite3 *db, sqlite3_stmt *stmt) {
     return OS_INVALID;
 }
 
-int wm_task_manager_check_db() {
-    const char *ROOT = ROOTUSER;
-    const char *path = TASKS_DB;
-    const char *sql;
-    const char *tail;
-    sqlite3 *db = NULL;
-    sqlite3_stmt *stmt = NULL;
-
-    w_mutex_lock(&db_mutex);
-
-    // Open or create the database file
-    if (sqlite3_open_v2(path, &db, SQLITE_OPEN_READWRITE | SQLITE_OPEN_CREATE, NULL)) {
-        mterror(WM_TASK_MANAGER_LOGTAG, MOD_TASK_CREATE_DB_ERROR);
-        w_mutex_unlock(&db_mutex);
-        return wm_task_manager_sql_error(db, stmt);
-    }
-
-    // Load the tables schema
-    for (sql = schema_task_manager_sql; sql && *sql; sql = tail) {
-        if (wdb_prepare(db, sql, -1, &stmt, &tail) != SQLITE_OK) {
-            mterror(WM_TASK_MANAGER_LOGTAG, MOD_TASK_SQL_PREPARE_ERROR);
-            w_mutex_unlock(&db_mutex);
-            return wm_task_manager_sql_error(db, stmt);
-        }
-
-        switch (wdb_step(stmt)) {
-        case SQLITE_MISUSE:
-        case SQLITE_ROW:
-        case SQLITE_DONE:
-        case SQLITE_CONSTRAINT:
-            break;
-        default:
-            mterror(WM_TASK_MANAGER_LOGTAG, MOD_TASK_SQL_STEP_ERROR);
-            w_mutex_unlock(&db_mutex);
-            return wm_task_manager_sql_error(db, stmt);
-        }
-
-        wdb_finalize(stmt);
-    }
-
-    sqlite3_close_v2(db);
-
-    w_mutex_unlock(&db_mutex);
-
-    uid_t uid = Privsep_GetUser(ROOT);
-    gid_t gid = Privsep_GetGroup(GROUPGLOBAL);
-
-    if (uid == (uid_t) - 1 || gid == (gid_t) - 1) {
-        mterror(WM_TASK_MANAGER_LOGTAG, USER_ERROR, ROOT, GROUPGLOBAL, strerror(errno), errno);
-        return OS_INVALID;
-    }
-
-    if (chown(path, uid, gid) < 0) {
-        mterror(WM_TASK_MANAGER_LOGTAG, CHOWN_ERROR, path, errno, strerror(errno));
-        return OS_INVALID;
-    }
-
-    if (chmod(path, 0660) < 0) {
-        mterror(WM_TASK_MANAGER_LOGTAG, CHMOD_ERROR, path, errno, strerror(errno));
-        return OS_INVALID;
-    }
-
-    return 0;
-}
-
-void* wm_task_manager_clean_db(void *arg) {
-    wm_task_manager *config = (wm_task_manager *)arg;
-    time_t next_clean = time(0);
-    time_t next_timeout = time(0);
-
-    while (1) {
-        time_t now = time(0);
-        time_t sleep_time = 0;
-
-        if (now >= next_timeout) {
-            // Set the status of old tasks IN PROGRESS to TIMEOUT
-            next_timeout = now + config->task_timeout;
-            wm_task_manager_set_timeout_status(now, config->task_timeout, &next_timeout);
-        }
-
-        if (now >= next_clean) {
-            // Delete entries older than cleanup_time
-            next_clean = now + WM_TASK_CLEANUP_DB_SLEEP_TIME;
-            wm_task_manager_delete_old_entries((now - config->cleanup_time));
-        }
-
-        if (next_timeout < next_clean) {
-            sleep_time = next_timeout;
-        } else {
-            sleep_time = next_clean;
-        }
-
-        w_sleep_until(sleep_time);
-
-    #ifdef WAZUH_UNIT_TESTING
-        break;
-    #endif
-    }
-
-    return NULL;
-}
-
-STATIC int wm_task_manager_set_timeout_status(time_t now, int timeout, time_t *next_timeout) {
+int wm_task_manager_set_timeout_status(time_t now, int timeout, time_t *next_timeout) {
     sqlite3 *db = NULL;
     sqlite3_stmt *stmt = NULL;
     sqlite3_stmt *stmt2 = NULL;
@@ -228,7 +109,7 @@ STATIC int wm_task_manager_set_timeout_status(time_t now, int timeout, time_t *n
     return OS_SUCCESS;
 }
 
-STATIC int wm_task_manager_delete_old_entries(int timestamp) {
+int wm_task_manager_delete_old_entries(int timestamp) {
     sqlite3 *db = NULL;
     sqlite3_stmt *stmt = NULL;
     int result = OS_INVALID;
@@ -576,58 +457,6 @@ int wm_task_manager_get_upgrade_task_by_agent_id(int agent_id, char **node, char
         sqlite_strdup((char*)sqlite3_column_text(stmt, 7), *status);
         sqlite_strdup((char*)sqlite3_column_text(stmt, 8), *error);
         result = task_id;
-    }
-
-    wdb_finalize(stmt);
-
-    sqlite3_close_v2(db);
-
-    w_mutex_unlock(&db_mutex);
-
-    return result;
-}
-
-int wm_task_manager_get_task_by_task_id(int task_id, char **node, char **module, char **command, char **status, char **error, int *create_time, int *last_update_time) {
-    sqlite3 *db = NULL;
-    sqlite3_stmt *stmt = NULL;
-    int result = OS_INVALID;
-    int agent_id;
-
-    w_mutex_lock(&db_mutex);
-
-    if (sqlite3_open_v2(TASKS_DB, &db, SQLITE_OPEN_READONLY, NULL)) {
-        mterror(WM_TASK_MANAGER_LOGTAG, MOD_TASK_OPEN_DB_ERROR);
-        w_mutex_unlock(&db_mutex);
-        return wm_task_manager_sql_error(db, stmt);
-    }
-
-    if (wdb_prepare(db, task_queries[WM_TASK_GET_TASK_BY_TASK_ID], -1, &stmt, NULL) != SQLITE_OK) {
-        mterror(WM_TASK_MANAGER_LOGTAG, MOD_TASK_SQL_PREPARE_ERROR);
-        w_mutex_unlock(&db_mutex);
-        return wm_task_manager_sql_error(db, stmt);
-    }
-
-    sqlite3_bind_int(stmt, 1, task_id);
-
-    switch (wdb_step(stmt)) {
-    case SQLITE_ROW:
-        agent_id = sqlite3_column_int(stmt, 1);
-        sqlite_strdup((char*)sqlite3_column_text(stmt, 2), *node);
-        sqlite_strdup((char*)sqlite3_column_text(stmt, 3), *module);
-        sqlite_strdup((char*)sqlite3_column_text(stmt, 4), *command);
-        *create_time = sqlite3_column_int(stmt, 5);
-        *last_update_time = sqlite3_column_int(stmt, 6);
-        sqlite_strdup((char*)sqlite3_column_text(stmt, 7), *status);
-        sqlite_strdup((char*)sqlite3_column_text(stmt, 8), *error);
-        result = agent_id;
-        break;
-    case SQLITE_DONE:
-        result = OS_NOTFOUND;
-        break;
-    default:
-        mterror(WM_TASK_MANAGER_LOGTAG, MOD_TASK_SQL_STEP_ERROR);
-        w_mutex_unlock(&db_mutex);
-        return wm_task_manager_sql_error(db, stmt);
     }
 
     wdb_finalize(stmt);
