+++ src/wazuh_modules/task_manager/wm_task_manager_db.h
@@ -25,7 +25,6 @@ typedef enum _task_query {
     WM_TASK_GET_LAST_AGENT_TASK,
     WM_TASK_GET_LAST_AGENT_UPGRADE_TASK,
     WM_TASK_UPDATE_TASK_STATUS,
-    WM_TASK_GET_TASK_BY_TASK_ID,
     WM_TASK_GET_TASK_BY_STATUS,
     WM_TASK_DELETE_OLD_TASKS,
     WM_TASK_DELETE_TASK,
@@ -35,17 +34,20 @@ typedef enum _task_query {
 extern char *schema_task_manager_sql;
 
 /**
- * Create the tasks DB or check that it already exists.
- * @return 0 when succeed, -1 otherwise.
+ * Update old tasks with status in progress to status timeout
+ * @param now Actual time
+ * @param timeout Task timeout
+ * @param next_timeout Next task in progress timeout
+ * @return OS_SUCCESS on success, OS_INVALID on errors
  * */
-int wm_task_manager_check_db();
+int wm_task_manager_set_timeout_status(time_t now, int timeout, time_t *next_timeout) __attribute__((nonnull));
 
 /**
- * Set tasks status to TIMEOUT after they are IN PROGRESS for a long period of time.
- * Delete entries older than a configurable period of time from the tasks DB.
- * @param arg Module configuration.
+ * Delete old tasks from the tasks DB
+ * @param timestamp Deletion limit time
+ * @return OS_SUCCESS on success, OS_INVALID on errors
  * */
-void* wm_task_manager_clean_db(void *arg) __attribute__((nonnull));
+int wm_task_manager_delete_old_entries(int timestamp);
 
 /**
  * Insert a new task in the tasks DB.
@@ -96,23 +98,6 @@ int wm_task_manager_cancel_upgrade_tasks(const char *node) __attribute__((nonnul
  * @return task_id when succeed, < 0 otherwise.
  * */
 int wm_task_manager_get_upgrade_task_by_agent_id(int agent_id, char **node, char **module, char **command, char **status, char **error, int *create_time, int *last_update_time) __attribute__((nonnull));
-
-/**
- * Get task by task_id from the tasks DB.
- * @param task_id ID of the task_id where the task is being executed.
- * @param node Node that executed the command.
- * @param module Name of the module where the command comes from.
- * @param command String where the command of the task will be stored.
- * @param status String where the status of the task will be stored.
- * @param error String where the error message of the task will be stored.
- * @param create_time Integer where the create_time of the task will be stored.
- * @param last_update_time Integer where the last_update_time of the task will be stored.
- * @return result
- * @retval OS_INVALID on errors
- * @retval agent_id if the task was found
- * @retval OS_NOTFOUND if the task was not found
- * */
-int wm_task_manager_get_task_by_task_id(int task_id, char **node, char **module, char **command, char **status, char **error, int *create_time, int *last_update_time) __attribute__((nonnull));
 
 #endif
 #endif
